feat: enhance Forge/NeoForge installer integration and progress reporting
- Refactor error handling in `create_zip_from_dir` and `copy_dir_recursive` to use `std::io::Error::other`.
- Add new functions to handle installer output and progress reporting during installation.
- Introduce `minecraft_version_from_installer` to extract Minecraft version from installer jars.
- Update `DetectedServerInfo` to include `loader_installer` and `needs_loader_install` fields.
- Modify `Instance` structure to accommodate new launch modes, including "installer".
- Implement progress tracking for the Forge/NeoForge installation process in the UI.
- Enhance user feedback during server import and installation processes with toast notifications.
- Add tests for new functionality related to installer version extraction and progress reporting.

+++ src-tauri/src/server/alerts.rs
@@ -86,7 +86,7 @@ async fn tick(app: &AppHandle) -> Result<(), String> {
         if max_ram_mb <= 0 {
             continue;
         }
-        let is_high = (usage.memory_mb as f64) >= (max_ram_mb as f64) * MEMORY_HIGH_FRACTION;
+        let is_high = usage.memory_mb >= (max_ram_mb as f64) * MEMORY_HIGH_FRACTION;
 
         let should_alert = {
             let state = app.state::<AppState>();
